Develop

+++ shadowsocks-csharp/Model/Server.cs
@@ -107,8 +107,8 @@ namespace Shadowsocks.Model
     {
         public string id;
         public string server;
-        public int server_port;
-        public int server_udp_port;
+        public ushort server_port;
+        public ushort server_udp_port;
         public string password;
         public string method;
         public string protocol;
@@ -382,7 +382,7 @@ namespace Shadowsocks.Model
                 throw new FormatException();
 
             server = match.Groups[1].Value;
-            server_port = int.Parse(match.Groups[2].Value);
+            server_port = ushort.Parse(match.Groups[2].Value);
             protocol = match.Groups[3].Value.Length == 0 ? "origin" : match.Groups[3].Value;
             protocol = protocol.Replace("_compatible", "");
             method = match.Groups[4].Value;
@@ -414,7 +414,7 @@ namespace Shadowsocks.Model
             }
             if (params_dict.ContainsKey("udpport"))
             {
-                server_udp_port = int.Parse(params_dict["udpport"]);
+                server_udp_port = ushort.Parse(params_dict["udpport"]);
             }
             if (!String.IsNullOrEmpty(force_group))
                 group = force_group;
@@ -437,7 +437,7 @@ namespace Shadowsocks.Model
             method = match.Groups["method"].Value;
             password = match.Groups["password"].Value;
             server = match.Groups["hostname"].Value;
-            server_port = int.Parse(match.Groups["port"].Value);
+            server_port = ushort.Parse(match.Groups["port"].Value);
             if (!String.IsNullOrEmpty(force_group))
                 group = force_group;
             else
